CRUD de convocatorias,usuarios,mentorias y htmls (frontend admin)

+++ src/main/java/com/usta/proyectointegrador/entities/MentoriaDTO.java
@@ -2,11 +2,11 @@ package com.usta.proyectointegrador.entities;
 
 public class MentoriaDTO {
 
-    private Integer idTransaccion;
+    private Long idTransaccion;
     private String nombreMentor;
     private String nombreStartup;
 
-    public MentoriaDTO(Integer idTransaccion,String nombreMentor, String nombreStartup) {
+    public MentoriaDTO(Long idTransaccion,String nombreMentor, String nombreStartup) {
         this.idTransaccion = idTransaccion;
         this.nombreMentor = nombreMentor;
         this.nombreStartup = nombreStartup;
@@ -23,10 +23,10 @@ public class MentoriaDTO {
     public void setNombreStartup(String nombreStartup) {
         this.nombreStartup = nombreStartup;
     }
-    public Integer getIdTransaccion() {
+    public Long getIdTransaccion() {
         return idTransaccion;
     }
-    public void setIdTransaccion(Integer idTransaccion) {
+    public void setIdTransaccion(Long idTransaccion) {
         this.idTransaccion = idTransaccion;
     }
